Fix timestamp mismatch

+++ src/main/java/cn/edu/nju/se/npmdependency/vo/RankUnitVO.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class RankUnitVO {
     String packageName;
     String version;
-    Integer count;
+    Long count;
 }
